+++ tests/testthat/test-bm_trim.R
@@ -1,0 +1,23 @@
+test_that("bm_trim()", {
+    font_file <- system.file("fonts/spleen/spleen-8x16.hex.gz", package = "bittermelon")
+    font <- read_hex(font_file)
+    capital_r <- font[[str2ucp("R")]]
+    capital_r_trimmed <- bm_trim(capital_r, c(1, 1, 3, 0))
+    expect_equal(nrow(capital_r_trimmed), 12L)
+    expect_equal(ncol(capital_r_trimmed), 7L)
+    verify_output("txt/capital_r_trimmed.txt",
+                  print(capital_r_trimmed, labels = c("-", "#")))
+
+    verify_output("txt/capital_r_trimmed_top.txt",
+                  print(bm_trim(capital_r, height = 14, vjust = "top"),
+                        labels = c("-", "#")))
+    verify_output("txt/capital_r_trimmed_bottom.txt",
+                  print(bm_trim(capital_r, height = 14, vjust = "bottom"),
+                        labels = c("-", "#")))
+    verify_output("txt/capital_r_trimmed_left.txt",
+                  print(bm_trim(capital_r, width = 6, hjust = "left"),
+                        labels = c("-", "#")))
+    verify_output("txt/capital_r_trimmed_right.txt",
+                  print(bm_trim(capital_r, width = 6, hjust = "right"),
+                        labels = c("-", "#")))
+})